+++ HISTORY.md
@@ -43,6 +43,9 @@
 * **Translation Updates:**
     * German
 
+### 2022-04-26
+* gui: distinguish between embedded blocks code and raw data in PNGs
+
 ### 2022-04-25
 * morphic, gui, objects, extensions: renamed "embeddedCode" property of costumes to "embeddedData"